requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
reportlab==4.3.1
drf-yasg==1.21.8
weasyprint==60.1
